enemyに対してAttackTypeがballetobjectの呼び出し要請を出し、受け取ったAMManagerがList配列で呼び出す処理を追加 bulletとenemyが衝突した際の処理を追加

+++ Scenes/PlayScene/UI/MachineSelect.cpp
@@ -81,8 +81,8 @@ void MachineSelect::DisplayObject(Microsoft::WRL::ComPtr<ID3D11ShaderResourceVie
 
 	// ƒ[ƒ‹ƒhÀ•W•ÔŠÒ
 	DirectX::SimpleMath::Vector3 worldPos = CalcScreenToXZN(
-		m_data.pos.x,
-		m_data.pos.y,
+		(int)m_data.pos.x,
+		(int)m_data.pos.y,
 		(int)pDR->GetOutputSize().right,
 		(int)pDR->GetOutputSize().bottom,
 		view,
